Fix tests to work on modern Rakudo
Somewhere along the line, turning a DateTime object into a Str changed
formats. I updated this format, using the validation pattern for
matching the ISO-8601-style timestamps used in Rakudo.

I also reformatted the regexes for matching to be clearer, and used a
variable for referencing line numbers rather than hardcoding a line
number in the message.